requirements.txt
# fastapi==0.100.1
# uvicorn[standard]==0.23.0
# pydantic==1.10.12
# aiofiles==23.2.1
# redis==5.0.0
# celery==5.3.4
# pillow==10.0.0  # Update to a valid version
# requests==2.31.0
# psycopg2-binary 
# sqlalchemy 
# databases 
# asyncpg

fastapi
uvicorn
pydantic
sqlalchemy
asyncpg
httpx
pandas
aiofiles
pillow
python-multipart
aiohttp
google-cloud-storage


